Data for One SRCV and One MRCV from the Kansas Farmer Survey

Responses for one SRCV and one MRCV from a survey of Kansas farmers. This data was first examined by Loughin and Scherer (1998) and subsequently examined by papers such as Agresti and Liu (1999) and Bilder, Loughin, and Nettleton (2000).

farmer1

Format

The data frame contains the following 6 columns:

Column 1, labeled Ed, corresponds to the respondent's highest attained level of education. A total of r = 5 levels of education are represented in the data.

1: High school

2: Vocational school

3: 2-Year college

4: 4-Year college

5: Other

Columns 2-6 correspond to the response categories for the question "What are your primary sources of veterinary information?" Binary responses (1 = Yes, 0 = No) are provided for each category.

Y1: Professional consultant

Y2: Veterinarian

Y3: State or local extension service

Y4: Magazines

Y5: Feed companies and representatives

Source

Loughin, T. M. and Scherer, P. N. (1998) Testing for association in contingency tables with multiple column responses. Biometrics, 54, 630–637.

References

Agresti, A. and Liu, I.-M. (1999) Modeling a categorical variable allowing arbitrarily many category choices. Biometrics, 55, 936–943.

Bilder, C., Loughin, T., and Nettleton, D. (2000) Multiple marginal independence testing for pick any/c variables. Communications in Statistics–Theory and Methods, 29, 1285–1316.

Loughin, T. M. and Scherer, P. N. (1998) Testing for association in contingency tables with multiple column responses. Biometrics, 54, 630–637.