Data for Three MRCVs from the Kansas Farmer Survey

Responses for three MRCVs from a survey of Kansas farmers. This data was first examined by Bilder and Loughin (2007).

Format

The data frame contains the following 12 columns:

Columns 1-3 correspond to the response categories for the question "Which of the following do you test your swine waste for?" Binary responses (1 = Yes, 0 = No) are provided for each category.

W1: Nitrogen

W2: Phosphorous

W3: Salt

Columns 4-7 correspond to the response categories for the question "What swine waste disposal methods do you use?" Binary responses (1 = Yes, 0 = No) are provided for each category.

Y1: Lagoon

Y2: Pit

Y3: Natural drainage

Y4: Holding tank

Columns 8-12 correspond to the response categories for the question "What are your primary sources of veterinary information?" Binary responses (1 = Yes, 0 = No) are provided for each category.

Z1: Professional consultant

Z2: Veterinarian

Z3: State or local extension service

Z4: Magazines

Z5: Feed companies and representatives

Source

Bilder, C. and Loughin, T. (2007) Modeling association between two or more categorical variables that allow for multiple category choices. Communications in Statistics–Theory and Methods, 36, 433–451.

Examples

data(farmer3)