Calculate Observed and Model-Predicted Odds Ratios for MRCV Data

The predict.genloglin method calculates observed and model-predicted odds ratios and their confidence intervals using results from genloglin. It uses an asymptotic normal approximation for the confidence intervals for both the observed and model-predicted odds ratios, and provides a bootstrap approach for estimating the confidence intervals of the model-predicted odds ratios.

Usage

# S3 method for class 'genloglin'
predict(object, alpha = 0.05, pair = "WY", print.status = TRUE, ...)

Arguments

object

An object of class "genloglin", produced by the genloglin function.

alpha

A numeric value specifying the desired alpha level for the confidence intervals. The function provides two-sided (1-alpha)x100% confidence intervals.

pair

For the case of three MRCVs, a character string specifying the pair of items for which odds ratios will be calculated: "WY" for (Wi, Yj) conditional on Zk, "WZ" for (Wi, Zk) conditional on Yj, and "YZ" for (Yj, Zk) conditional on Wi.

print.status

A logical value indicating whether to provide bootstrap progress updates.

...

Additional arguments passed to or from other methods.

Value

A list containing at least original.arg, OR.obs, and OR.model.asymp:

original.arg: A list containing the original arguments supplied to genloglin.

OR.obs: A numeric matrix containing observed odds ratios and their confidence intervals.

OR.model.asymp: A numeric matrix containing model-predicted odds ratios and their confidence intervals using asymptotic normal approximation.

For boot = TRUE in the genloglin function, the returned list additionally includes:

boot.results: A list containing bootstrap-related results such as B.use, B.discard, and OR.model.BCa.

Details

Wald confidence intervals are estimated for both model-based (see Appendix A of Bilder and Loughin, 2007) and observed (see Agresti, 2013, p. 70) odds ratios.

A bootstrap method is also available, providing bias-corrected accelerated (BCa) confidence intervals for the model-predicted odds ratios. For more information on BCa intervals, see Efron (1987). The predict.genloglin function uses a jackknife approximation to estimate empirical influence values.

Note: The bootstrap confidence intervals are only available when boot = TRUE in the original call to the genloglin function.

References

Agresti, A. (2013). Categorical Data Analysis (3rd ed.). Hoboken, NJ: John Wiley & Sons.

Bilder, C., & Loughin, T. (2007). Modeling association between two or more categorical variables that allow for multiple category choices. Communications in Statistics–Theory and Methods, 36, 433-451.

Efron, B. (1987). Better bootstrap confidence intervals. Journal of the American Statistical Association, 82, 171-185.

Examples

## For examples see help(genloglin).