Control Printed Display of Objects of Class "MMI"

The print.MMI method function controls the printed display of objects of class MMI.

The print.SPMI method function controls the printed display of objects of class SPMI.

A function used to control the display of output provided by anova.genloglin.

Method functions that control the printed display of MRCV regression modeling objects.

A function used to control the display of output provided by predict.genloglin.

A method function that controls the display of summary.genloglin

Usage

# S3 method for class 'MMI'
print(x, ...)

# S3 method for class 'SPMI'
print(x, ...)

# S3 method for class 'anova.genloglin'
print(x, ...)

# S3 method for class 'genloglin'
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for class 'predict.genloglin'
print(x, ...)

# S3 method for class 'summary.genloglin'
print(
  x,
  digits = max(3, getOption("digits") - 3),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)

Arguments

x

An object of class "genloglin", produced by the genloglin function.

...

Additional arguments passed to or from other methods.

digits

Minimum number of digits; see print.default for additional explanation.

symbolic.cor

A logical value indicating whether correlations should be printed in symbolic form; see the summary method for class glm for additional explanation.

signif.stars

A logical value indicating whether significance stars should be printed; see the summary method for class glm for additional explanation.

Value

A print out of selected results from MI.test.

A print out of selected results from MI.test.

A print out of selected results.

Details

The print.genloglin function is based on the print method for class "glm".

The print.summary.genloglin function is based on the print method for class "summary.glm".