Control Printed Display of MRCV Regression Modeling Objects

Method functions that control the printed display of MRCV regression modeling objects.

# S3 method for class 'genloglin'
print(x, digits = max(3, getOption("digits") - 3), ...)
# S3 method for class 'summary.genloglin'
print(x, digits = max(3, getOption("digits") - 3), symbolic.cor = 
    x$symbolic.cor, signif.stars = getOption("show.signif.stars"), ...)
# S3 method for class 'anova.genloglin'
print(x, ...)
# S3 method for class 'predict.genloglin'
print(x, ...)

Arguments

x

An object of class "genloglin" produced by the genloglin function.

digits

Minimum number of digits; see print.default for additional explanation.

symbolic.cor

A logical value indicating whether correlations should be printed in symbolic form; see the summary method for class "glm" for additional explanation.

signif.stars

A logical value indicating whether significance stars should be printed; see the summary method for class "glm" for additional explanation.

...

Additional arguments passed to or from other methods.

Details

The print.genloglin function is based on the print method for class "glm".

The print.summary.genloglin function is based on the print method for class "summary.glm".

Value

A print out of selected results.