Calculate Standardized Pearson Residuals for MRCV Data

The residuals.genloglin method function calculates standardized Pearson residuals for the model specified in the genloglin function. It offers an asymptotic approximation and a bootstrap approximation for estimating the variance of the residuals.

# S3 method for class 'genloglin'
residuals(object, ...)

Arguments

object

An object of class 'genloglin' produced by the genloglin function.

...

Additional arguments passed to or from other methods.

Details

The bootstrap results are only available when boot = TRUE in the call to the genloglin function.

The residuals.genloglin function uses tables:tabular() to display the results for the two MRCV case.

See Bilder and Loughin (2007) for additional details about calculating the residuals.

Value

— A list containing at least std.pearson.res.asymp.var. For the two MRCV case, the object is a 2Ix2J table of class 'tabular' containing the standardized Pearson residuals based on the estimated asymptotic variance. For the three MRCV case, the object is a data frame containing the 2Ix2Jx2K residuals.

— For boot = TRUE in the call to the genloglin function, the list additionally includes:

B.use: The number of bootstrap resamples used.

B.discard: The number of bootstrap resamples discarded due to having at least one item with all positive or negative responses.

std.pearson.res.boot.var: For the two MRCV case, a 2Ix2J table of class 'tabular' containing the standardized Pearson residuals based on the bootstrap variance. For the three MRCV case, a data frame containing the 2Ix2Jx2K residuals.

References

Bilder, C. and Loughin, T. (2007) Modeling association between two or more categorical variables that allow for multiple category choices. Communications in Statistics–Theory and Methods, 36, 433–451.

Examples

## For examples see help(genloglin).