Steller Sea Lion Scat Data of Riemer, Wright, and Brown (2011)

The sealion data frame is a portion of the data analyzed by Riemer, Wright, and Brown (2011). The data represents the types of prey found in Steller sea lion scats found at the mouth of the Columbia river in August 2004 and 2007. The goal was to determine if the diet habits of the sea lions had changed over time.

sealion

Format

The data frame contains the following 12 columns:

Column 1, labeled Date, corresponds to the date of data collection. Two collection dates, 8/10/2004 and 8/7/2007, are included in the data frame.

Columns 2-12 correspond to the types of prey found in Steller sea lion scats. Binary responses (1 = Present in scat, 0 = Not present in scat) are provided for each category.

prey1: Unidentified fish

prey2: Pacific lamprey (Lampetra tridentata)

prey3: Starry flounder (Platichthys stellatus)

prey4: Pacific sardine (Sardinops sagax)

prey5: Pacific herring (Clupea pallasii)

prey6: Unidentified clupeid (family Clupeidae)

prey7: Unidentified skate (family Rajidae)

prey8: Northern anchovy (Engraulis mordax)

prey9: Pacific salmon (Oncorhynchus spp.)

prey10: Pacific staghorn sculpin (Leptocottus armatus)

prey11: Pacific hake (Merluccius productus)

Source

Riemer, S. D., Wright, B. E., and Brown, R. F. (2011) Food habits of Steller sea lions (Eumetopias jubatus) off Oregon and northern California, 1986-2007. Fishery Bulletin, 109, 369–381.